isolate disc code from libretro.cpp; support eject/close tray
