+++ Makefile.common
@@ -210,6 +210,7 @@ SOURCES_CXX += \
 	$(MEDNAFEN_DIR)/video/Deinterlacer.cpp \
 	$(MEDNAFEN_DIR)/video/surface.cpp \
 	$(MEDNAFEN_DIR)/git.cpp \
+	$(CORE_DIR)/disc.cpp \
 	$(CORE_DIR)/input.cpp \
 	$(CORE_DIR)/libretro.cpp \
 	$(CORE_DIR)/libretro_settings.cpp
