+++ libretro_settings.cpp
@@ -9,5 +9,5 @@ int setting_initial_scanline_pal = 0;
 int setting_last_scanline = 239;
 int setting_last_scanline_pal = 287;
 int setting_gun_crosshair = SETTING_GUN_CROSSHAIR_CROSS;
-
+bool setting_disc_test = false;
