+++ libretro_settings.h
@@ -1,6 +1,15 @@
 #ifndef __LIBRETRO_SETTINGS_HDR__
 #define __LIBRETRO_SETTINGS_HDR__
 
+enum
+{
+	SETTING_GUN_CROSSHAIR_OFF,
+	SETTING_GUN_CROSSHAIR_CROSS,
+	SETTING_GUN_CROSSHAIR_DOT,
+
+	SETTING_GUN_CROSSHAIR_LAST,
+};
+
 extern int setting_region;
 extern int setting_cart;
 extern bool setting_smpc_autortc;
@@ -9,15 +18,7 @@ extern int setting_initial_scanline;
 extern int setting_initial_scanline_pal;
 extern int setting_last_scanline;
 extern int setting_last_scanline_pal;
-
-enum {
-	SETTING_GUN_CROSSHAIR_OFF,
-	SETTING_GUN_CROSSHAIR_CROSS,
-	SETTING_GUN_CROSSHAIR_DOT,
-
-	SETTING_GUN_CROSSHAIR_LAST,
-};
-
 extern int setting_gun_crosshair;
+extern bool setting_disc_test;
 
 #endif
